+++ meta_package_manager/__main__.py
@@ -34,11 +34,11 @@ def main():
     <https://github.com/python-poetry/poetry/issues/5981>`_:
 
         - plain inline package call: ``python -m meta_package_manager``,
-        - Poetry's script entry point: ``mpm = 'meta_package_manager.__main__:main``,
-        - Nuitka's main module invocation requirement: ``python -m nuitka (...)
-          meta_package_manager/__main__.py``
+        - ``pyproject.toml`` entry point: ``mpm = 'meta_package_manager.__main__:main``,
+        - Nuitka's main module invocation requirement:
+          ``python -m nuitka (...) meta_package_manager/__main__.py``
 
-    That way we can deduce all three cases from the Poetry entry point.
+    That way we can deduce all three cases from the entry point.
     """
     from meta_package_manager.cli import mpm